new addSession component on its own file, cleaned up css/random js not in use
