+++ src/App.scss
@@ -1,6 +1,3 @@
-body{
-
-}
 .App {
   text-align: center;
   background-color:white;
